[poincare, apps] Reduce the number of run-time float <-> double conversions, factor out methods between the FloatPairStore children, perform a bit of manual CSE.
This reduces accuracy a tiny bit, but... if accuracy were a goal, only double would be used in the first place.

+++ apps/probability/law/normal_law.cpp
@@ -68,7 +68,7 @@ float NormalLaw::evaluateAtAbscissa(float x) const {
   if (m_parameter2 == 0.0f) {
     return NAN;
   }
-  return (1.0f/(std::fabs(m_parameter2)*std::sqrt(2.0f*M_PI)))*std::exp(-0.5f*std::pow((x-m_parameter1)/m_parameter2,2));
+  return (1.0f/(std::fabs(m_parameter2)*std::sqrt(2.0f*(float)M_PI)))*std::exp(-0.5f*std::pow((x-m_parameter1)/m_parameter2,2));
 }
 
 bool NormalLaw::authorizedValueAtIndex(float x, int index) const {
